examples/vector_embeddings/pyproject.toml
[tool.poetry]
name = "vector-embeddings"
version = "0.1.0"
description = ""
authors = ["Jake Zuliani <jake@burla.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-compute = "^1.33.0"
google-auth = "^2.40.3"
google-cloud-storage = "^3.2.0"
burla = "^1.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
